start storing exp.scripts; refit batch fab to comprehension
Hey, it's cool to store experiment scripts! The *reasonable* way is to make a branch for *each* experiment I think, but I am just checking for now, so all's in `t-soplib06` branch. Then the flow could be (1)build dpm (2)run the `./exp-script/exp.py` (3) execute the `batch.csv` it produces.
In addition, I refitted the batch fab script into a neat *list comprehension*. Might also set all the parameters WITHIN the list, but we'll see what is better.